refactor: SlidingPiece 추상화
