+++ src/main/java/domain/piece/Bishop.java
@@ -1,13 +1,11 @@
 package domain.piece;
 
 import domain.Direction;
-import domain.Square;
 import domain.Team;
 
 import java.util.List;
-import java.util.Map;
 
-public class Bishop extends Piece {
+public class Bishop extends SlidingPiece {
     public Bishop(final Team team) {
         super(team);
     }
@@ -19,11 +17,5 @@ public class Bishop extends Piece {
                 Direction.NORTH_WEST,
                 Direction.SOUTH_EAST,
                 Direction.SOUTH_WEST);
-    }
-
-    @Override
-    public boolean canMove(final Square source, final Square target, final Map<Square, Piece> pieces) {
-        final List<Square> path = calculatePath(source, pieces);
-        return path.contains(target);
     }
 }
