+++ src/main/java/domain/piece/Piece.java
@@ -4,7 +4,6 @@ import domain.Direction;
 import domain.Square;
 import domain.Team;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -16,28 +15,6 @@ public abstract class Piece {
     }
 
     protected abstract List<Direction> movableDirections();
-
-    protected List<Square> calculatePath(final Square source, final Map<Square, Piece> pieces) {
-        return movableDirections().stream()
-                .flatMap(direction -> calculatePathInDirection(source, direction, pieces).stream())
-                .toList();
-    }
-
-    private List<Square> calculatePathInDirection(final Square source, final Direction direction, final Map<Square, Piece> pieces) {
-        final List<Square> movableSquares = new ArrayList<>();
-        Square movableSource = source;
-
-        while (movableSource.canMove(direction)) {
-            movableSource = movableSource.next(direction);
-            if (pieces.containsKey(movableSource)) {
-                break;
-            }
-            movableSquares.add(movableSource);
-        }
-
-        return movableSquares;
-    }
-
 
     public boolean canNotMove(final Square source, final Square target, final Map<Square, Piece> pieces) {
         return !canMove(source, target, pieces);
