+++ src/main/java/domain/piece/Rook.java
@@ -1,13 +1,11 @@
 package domain.piece;
 
 import domain.Direction;
-import domain.Square;
 import domain.Team;
 
 import java.util.List;
-import java.util.Map;
 
-public class Rook extends Piece {
+public class Rook extends SlidingPiece {
     public Rook(final Team team) {
         super(team);
     }
@@ -19,11 +17,5 @@ public class Rook extends Piece {
                 Direction.SOUTH,
                 Direction.EAST,
                 Direction.WEST);
-    }
-
-    @Override
-    public boolean canMove(final Square source, final Square target, final Map<Square, Piece> pieces) {
-        final List<Square> path = calculatePath(source, pieces);
-        return path.contains(target);
     }
 }
